* Rename Metadata.GenericContext to MetadataGenericContext to avoid conflicts with TypeSystem.GenericContext * Change MetadataGenericContext to readonly struct * Rearrange types in Dom.cs
