+++ ICSharpCode.Decompiler/Disassembler/ILStructure.cs
@@ -60,7 +60,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 	{
 		public readonly PEFile Module;
 		public readonly MethodDefinitionHandle MethodHandle;
-		public readonly GenericContext GenericContext;
+		public readonly MetadataGenericContext GenericContext;
 		public readonly ILStructureType Type;
 
 		/// <summary>
@@ -88,7 +88,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 		/// </summary>
 		public readonly List<ILStructure> Children = new List<ILStructure>();
 
-		public ILStructure(PEFile module, MethodDefinitionHandle handle, GenericContext genericContext, MethodBodyBlock body)
+		public ILStructure(PEFile module, MethodDefinitionHandle handle, MetadataGenericContext genericContext, MethodBodyBlock body)
 			: this(module, handle, genericContext, ILStructureType.Root, 0, body.GetILReader().Length)
 		{
 			// Build the tree of exception structures:
@@ -142,7 +142,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			SortChildren();
 		}
 
-		public ILStructure(PEFile module, MethodDefinitionHandle handle, GenericContext genericContext, ILStructureType type, int startOffset, int endOffset, ExceptionRegion handler = default)
+		public ILStructure(PEFile module, MethodDefinitionHandle handle, MetadataGenericContext genericContext, ILStructureType type, int startOffset, int endOffset, ExceptionRegion handler = default)
 		{
 			Debug.Assert(startOffset < endOffset);
 			this.Module = module;
@@ -154,7 +154,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			this.ExceptionHandler = handler;
 		}
 
-		public ILStructure(PEFile module, MethodDefinitionHandle handle, GenericContext genericContext, ILStructureType type, int startOffset, int endOffset, int loopEntryPoint)
+		public ILStructure(PEFile module, MethodDefinitionHandle handle, MetadataGenericContext genericContext, ILStructureType type, int startOffset, int endOffset, int loopEntryPoint)
 		{
 			Debug.Assert(startOffset < endOffset);
 			this.Module = module;
