+++ ICSharpCode.Decompiler/Disassembler/ReflectionDisassembler.cs
@@ -143,7 +143,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 
 		public void DisassembleMethod(PEFile module, MethodDefinitionHandle handle)
 		{
-			var genericContext = new GenericContext(handle, module);
+			var genericContext = new MetadataGenericContext(handle, module);
 			// write method header
 			output.WriteReference(module, handle, ".method", isDefinition: true);
 			output.Write(" ");
@@ -153,14 +153,14 @@ namespace ICSharpCode.Decompiler.Disassembler
 
 		public void DisassembleMethodHeader(PEFile module, MethodDefinitionHandle handle)
 		{
-			var genericContext = new GenericContext(handle, module);
+			var genericContext = new MetadataGenericContext(handle, module);
 			// write method header
 			output.WriteReference(module, handle, ".method", isDefinition: true);
 			output.Write(" ");
 			DisassembleMethodHeaderInternal(module, handle, genericContext);
 		}
 
-		void DisassembleMethodHeaderInternal(PEFile module, MethodDefinitionHandle handle, GenericContext genericContext)
+		void DisassembleMethodHeaderInternal(PEFile module, MethodDefinitionHandle handle, MetadataGenericContext genericContext)
 		{
 			var metadata = module.Metadata;
 
@@ -343,7 +343,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 		}
 
 		void DisassembleMethodBlock(PEFile module, MethodDefinitionHandle handle,
-			GenericContext genericContext)
+			MetadataGenericContext genericContext)
 		{
 			var metadata = module.Metadata;
 			var methodDefinition = metadata.GetMethodDefinition(handle);
@@ -1132,7 +1132,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			output.WriteLine();
 		}
 
-		void WriteGenericParameterAttributes(PEFile module, GenericContext context, GenericParameterHandle handle)
+		void WriteGenericParameterAttributes(PEFile module, MetadataGenericContext context, GenericParameterHandle handle)
 		{
 			var metadata = module.Metadata;
 			var p = metadata.GetGenericParameter(handle);
@@ -1324,7 +1324,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			const FieldAttributes hasXAttributes = FieldAttributes.HasDefault | FieldAttributes.HasFieldMarshal | FieldAttributes.HasFieldRVA;
 			WriteFlags(fieldDefinition.Attributes & ~(FieldAttributes.FieldAccessMask | hasXAttributes), fieldAttributes);
 
-			var signature = fieldDefinition.DecodeSignature(new DisassemblerSignatureTypeProvider(module, output), new GenericContext(fieldDefinition.GetDeclaringType(), module));
+			var signature = fieldDefinition.DecodeSignature(new DisassemblerSignatureTypeProvider(module, output), new MetadataGenericContext(fieldDefinition.GetDeclaringType(), module));
 
 			var marshallingDescriptor = fieldDefinition.GetMarshallingDescriptor();
 			if (!marshallingDescriptor.IsNil)
@@ -1411,7 +1411,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			WriteFlags(propertyDefinition.Attributes, propertyAttributes);
 			var accessors = propertyDefinition.GetAccessors();
 			var declaringType = metadata.GetMethodDefinition(accessors.GetAny()).GetDeclaringType();
-			var signature = propertyDefinition.DecodeSignature(new DisassemblerSignatureTypeProvider(module, output), new GenericContext(declaringType, module));
+			var signature = propertyDefinition.DecodeSignature(new DisassemblerSignatureTypeProvider(module, output), new MetadataGenericContext(declaringType, module));
 
 			if (signature.Header.IsInstance)
 				output.Write("instance ");
@@ -1441,7 +1441,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 
 			output.Write(keyword);
 			output.Write(' ');
-			((EntityHandle)method).WriteTo(module, output, GenericContext.Empty);
+			((EntityHandle)method).WriteTo(module, output, default);
 			output.WriteLine();
 		}
 		#endregion
@@ -1506,7 +1506,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 					signature = provider.GetTypeFromReference(module.Metadata, (TypeReferenceHandle)eventDefinition.Type, 0);
 					break;
 				case HandleKind.TypeSpecification:
-					signature = provider.GetTypeFromSpecification(module.Metadata, new GenericContext(declaringType, module),
+					signature = provider.GetTypeFromSpecification(module.Metadata, new MetadataGenericContext(declaringType, module),
 						(TypeSpecificationHandle)eventDefinition.Type, 0);
 					break;
 				default:
@@ -1556,7 +1556,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 		public void DisassembleType(PEFile module, TypeDefinitionHandle type)
 		{
 			var typeDefinition = module.Metadata.GetTypeDefinition(type);
-			GenericContext genericContext = new GenericContext(type, module);
+			MetadataGenericContext genericContext = new MetadataGenericContext(type, module);
 
 			DisassembleTypeHeaderInternal(module, type, typeDefinition, genericContext);
 
@@ -1663,11 +1663,11 @@ namespace ICSharpCode.Decompiler.Disassembler
 		public void DisassembleTypeHeader(PEFile module, TypeDefinitionHandle type)
 		{
 			var typeDefinition = module.Metadata.GetTypeDefinition(type);
-			GenericContext genericContext = new GenericContext(type, module);
+			MetadataGenericContext genericContext = new MetadataGenericContext(type, module);
 			DisassembleTypeHeaderInternal(module, type, typeDefinition, genericContext);
 		}
 
-		private void DisassembleTypeHeaderInternal(PEFile module, TypeDefinitionHandle handle, TypeDefinition typeDefinition, GenericContext genericContext)
+		private void DisassembleTypeHeaderInternal(PEFile module, TypeDefinitionHandle handle, TypeDefinition typeDefinition, MetadataGenericContext genericContext)
 		{
 			output.WriteReference(module, handle, ".class", isDefinition: true);
 			WriteMetadataToken(output, module, handle, MetadataTokens.GetToken(handle),
@@ -1696,7 +1696,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			}
 		}
 
-		void WriteTypeParameters(ITextOutput output, PEFile module, GenericContext context, GenericParameterHandleCollection p)
+		void WriteTypeParameters(ITextOutput output, PEFile module, MetadataGenericContext context, GenericParameterHandleCollection p)
 		{
 			if (p.Count > 0)
 			{
@@ -1755,7 +1755,7 @@ namespace ICSharpCode.Decompiler.Disassembler
 			{
 				output.Write(".custom ");
 				var attr = metadata.GetCustomAttribute(a);
-				attr.Constructor.WriteTo(module, output, GenericContext.Empty);
+				attr.Constructor.WriteTo(module, output, default);
 				if (!attr.Value.IsNil)
 				{
 					output.Write(" = ");
