+++ ILSpy/Metadata/CorTables/ClassLayoutTableTreeNode.cs
@@ -112,7 +112,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string ParentTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new GenericContext(default(TypeDefinitionHandle), module);
+					var context = new MetadataGenericContext(default(TypeDefinitionHandle), module);
 					((EntityHandle)classLayout.Parent).WriteTo(module, output, context);
 					return output.ToString();
 				}
