+++ ILSpy/Metadata/CorTables/FieldLayoutTableTreeNode.cs
@@ -110,7 +110,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string FieldTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new Decompiler.Metadata.GenericContext(default(TypeDefinitionHandle), module);
+					var context = new Decompiler.Metadata.MetadataGenericContext(default(TypeDefinitionHandle), module);
 					((EntityHandle)fieldLayout.Field).WriteTo(module, output, context);
 					return output.ToString();
 				}
