+++ ILSpy/Metadata/CorTables/ImplMapTableTreeNode.cs
@@ -127,7 +127,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string MemberForwardedTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new GenericContext(default(TypeDefinitionHandle), module);
+					var context = new MetadataGenericContext(default(TypeDefinitionHandle), module);
 					((EntityHandle)implMap.MemberForwarded).WriteTo(module, output, context);
 					return output.ToString();
 				}
@@ -145,7 +145,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string ImportScopeTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new GenericContext(default(TypeDefinitionHandle), module);
+					var context = new MetadataGenericContext(default(TypeDefinitionHandle), module);
 					((EntityHandle)implMap.ImportScope).WriteTo(module, output, context);
 					return output.ToString();
 				}
