+++ ILSpy/Metadata/CorTables/MemberRefTableTreeNode.cs
@@ -99,7 +99,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string ParentTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					memberRef.Parent.WriteTo(module, output, Decompiler.Metadata.GenericContext.Empty);
+					memberRef.Parent.WriteTo(module, output, default);
 					return output.ToString();
 				}
 			}
@@ -114,7 +114,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string SignatureTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new Decompiler.Metadata.GenericContext(default(TypeDefinitionHandle), module);
+					var context = new MetadataGenericContext(default(TypeDefinitionHandle), module);
 					((EntityHandle)handle).WriteTo(module, output, context);
 					return output.ToString();
 				}
