+++ ILSpy/Metadata/CorTables/MethodTableTreeNode.cs
@@ -127,7 +127,7 @@ namespace ICSharpCode.ILSpy.Metadata
 					if (signatureTooltip == null)
 					{
 						ITextOutput output = new PlainTextOutput();
-						var context = new Decompiler.Metadata.GenericContext(default(TypeDefinitionHandle), module);
+						var context = new Decompiler.Metadata.MetadataGenericContext(default(TypeDefinitionHandle), module);
 						((EntityHandle)handle).WriteTo(module, output, context);
 						signatureTooltip = output.ToString();
 					}
