+++ ILSpy/Metadata/DebugTables/CustomDebugInformationTableTreeNode.cs
@@ -202,7 +202,7 @@ namespace ICSharpCode.ILSpy.Metadata
 			public string ParentTooltip {
 				get {
 					ITextOutput output = new PlainTextOutput();
-					var context = new GenericContext(default(TypeDefinitionHandle), module);
+					var context = new MetadataGenericContext(default(TypeDefinitionHandle), module);
 					debugInfo.Parent.WriteTo(module, output, context);
 					return output.ToString();
 				}
